refactor: present OpenShell support as experimental
