+++ docs/reference/api/rust-library-reference/nemo-fabric-core/error/enum-fabricerror.mdx
@@ -433,7 +433,7 @@ Runtime handle id.
 
 <pre className="rust-signature"><code dangerouslySetInnerHTML={{"__html": "UnsupportedEnvironmentProvider"}} /></pre>
 
-An environment provider is not runnable for the selected adapter in this POC.
+An environment provider is not runnable for the selected adapter.
 
 #### Fields
