+++ docs/sdk/python.mdx
@@ -562,7 +562,7 @@ it does not create the resource or transfer deletion authority. Runtime stop
 does not release either a prepared or attached environment. For a caller-owned
 handle, `release_environment(...)` detaches Fabric and leaves final deletion to
 the consumer. See the
-[LangGraph OpenShell vertical slice](https://github.com/NVIDIA/NeMo-Fabric/tree/main/examples/langgraph_openshell_poc)
+[LangGraph OpenShell example](https://github.com/NVIDIA/NeMo-Fabric/tree/main/examples/langgraph_openshell)
 for a runnable example of this deployment-oriented flow.
 
 ## Native OpenAI Streaming
