+++ examples/langgraph_openshell/adapter/__init__.py
