+++ examples/langgraph_openshell/consumer/config.py
