+++ examples/langgraph_openshell/runtime-image.dockerignore
